backend/booking_project/requirements.txt

asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.9
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.1.4
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
typing_extensions==4.12.2
whitenoise==6.7.0